Correct the behaviour with "octal or backreference" inside and outside char class
Never treat \ddd as a backreference inside a character class.

Inside character classes, \9 and \8 are now treated as '9' and '8',
respectively.

This must be refined when we add support for keeping track of groups
in scope of references, since it will affect how many digits will be
parsed, outside of character classes.